+++ backend-axum/src/models/message.rs
@@ -1,0 +1,14 @@
+use serde::{Deserialize, Serialize};
+use sqlx::FromRow;
+
+#[derive(Debug, Deserialize, Serialize, FromRow)]
+pub struct CreateMessageSchema {
+    pub message: String,
+}
+
+#[derive(Debug, Deserialize, Serialize, FromRow)]
+pub struct MessageModel {
+    pub id: i64,
+    pub message: String,
+    pub created_at: chrono::DateTime<chrono::Utc>,
+}
